Remove parameters remove_files and add_files
In my oppinion having a static list of files to be added or removed does
not make much sense for Concourse because pipelines are run multiple
times and after the commit there is no need to add or remove the same
file again because the package already reflects this change.

There are also cases where somebody during pipeline creation does not
know the file name which needed to be added to the package.

So instead of using this parameters I have added `osc addremove` to be
run all the time before building or committing the package.
This takes the content of the resource directory as a given, so removing
files is done the same way as adding files to the directory via a
Concourse task.